update triangle count type

+++ cpp/src/structure/edge_triangle_count_sg.cu
@@ -24,7 +24,7 @@ template rmm::device_uvector<int32_t> edge_triangle_count(
   raft::device_span<int32_t> edgelist_srcs,
   raft::device_span<int32_t> edgelist_dsts);
 
-template rmm::device_uvector<int32_t> edge_triangle_count(
+template rmm::device_uvector<int64_t> edge_triangle_count(
   raft::handle_t const& handle,
   cugraph::graph_view_t<int32_t, int64_t, false, false> const& graph_view,
   raft::device_span<int32_t> edgelist_srcs,
@@ -36,7 +36,7 @@ template rmm::device_uvector<int64_t> edge_triangle_count(
   raft::device_span<int64_t> edgelist_srcs,
   raft::device_span<int64_t> edgelist_dsts);
 
-template rmm::device_uvector<int64_t> edge_triangle_count(
+template rmm::device_uvector<int32_t> edge_triangle_count(
   raft::handle_t const& handle,
   cugraph::graph_view_t<int64_t, int32_t, false, false> const& graph_view,
   raft::device_span<int64_t> edgelist_srcs,
